Let Canvas2DLayerBridge support software rendering
Right now in Canvas2DLayerBridge::PrepareMailbox, we only support the
case when the rendering mode is GPU. This CL makes it support software
rendering mode. In particular, when creating the viz::TextureMailbox,
we create it from a SharedBitmap for the software rendering.

Bug: None
Change-Id: I698196cacf2a94f56623e79ea72c1a0bcd158d7a
Reviewed-on: https://chromium-review.googlesource.com/633484
WPT-Export-Revision: 1997d7ef32c96a0465c092d7248b8917b8b003ea